Training/requirements.txt
# Facial Recognition
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython>=3.1.20
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.0.98
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
opencv-python-headless

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Facenet --------------------------------------------------------------------
scikit-learn<=1.1.2  # CoreML quantization
tensorflow  # TF exports (-cpu, -aarch64, -macos)
h5py
imageio

# Deploy ----------------------------------------------------------------------
setuptools>=65.5.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
boto3
